GameObject.Find and GetComponent in one method.
A function for finding a GameObject, placing it in an existing reference, and then retrieving any one of its components and placing that inside another existing reference.

May be useful at runtime for finding and using prefabs/instantiated objects, and for retrieving their components for use. Use this method sparingly, ideally only on objects that are inaccessible during edit mode.